refactor: 리뷰 분석 결과 저장 위치 변경
데이터 정규화 및 중복 저장을 방지하기 위해 AI 분석 결과 필드를 CompetitorProduct 엔터티로 이동 (sentimentScore, reviewTags, positivePoints, painPoints)
