+++ backend/src/main/java/codebadger/virtual_launch/domain/crawling/domain/entity/RawReview.java
@@ -17,7 +17,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.hibernate.validator.constraints.Range;
 
 @Getter
 @Setter
@@ -43,25 +42,4 @@ public class RawReview { // 실제 리뷰 - 개별
     @Column(columnDefinition = "TEXT")
     private String originalContent; // 원본 리뷰 내용
     private int starRating; // 별점 (1~5)
-
-    @Column
-    @Range(min = 0, max = 10)
-    private double sentiment; // 긍정, 부정 스코어 점수 (0.0 ~ 10.0)
-
-    @Column(columnDefinition = "TEXT")
-    private String reviewTags; // 리뷰에서 추출된 키워드 (콤마로 구분)
-
-    @Column(columnDefinition = "TEXT")
-    private String positivePoints; // 핵심 만족 사항
-
-    @Column(columnDefinition = "TEXT")
-    private String painPoints; // 핵심 불만 사항
-
-    public void updateAnalysisResult(double sentiment, String reviewTags, String positivePoints, String painPoints) {
-        this.sentiment = sentiment;
-        this.reviewTags = reviewTags;
-        this.positivePoints = positivePoints;
-        this.painPoints = painPoints;
-
-    }
 }
